jdbc/src/main/java/io/simpolor/jdbc/controller/StudentController.java
package io.simpolor.jdbc.controller;

import io.simpolor.jdbc.model.StudentDto;
import io.simpolor.jdbc.repository.entity.Student;
import io.simpolor.jdbc.service.StudentService;
import lombok.RequiredArgsConstructor;
import org.springframework.util.CollectionUtils;
import org.springframework.web.bind.annotation.*;

import java.util.Collections;
import java.util.List;
import java.util.Objects;

@RestController
@RequestMapping("/students")
@RequiredArgsConstructor
public class StudentController {

	private final StudentService studentService;

	@GetMapping
	public List<StudentDto> list() {

		List<Student> students = studentService.getAll();
		if(CollectionUtils.isEmpty(students)){
			return Collections.EMPTY_LIST;
		}

		return StudentDto.of(students);
	}

	@GetMapping(value="/{studentId}")
	public StudentDto detail(@PathVariable Long studentId) {

		Student student = studentService.get(studentId);
		if(Objects.isNull(student)){
			return null;
		}

		return StudentDto.of(student);
	}

	@PostMapping
	public Long register(@RequestBody StudentDto studentDto) {

		return studentService.create(studentDto.toEntity());
	}

	@PutMapping(value="/{studentId}")
	public void modify(@PathVariable Long studentId,
					   @RequestBody StudentDto studentDto) {

		studentDto.setId(studentId);
		studentService.update(studentDto.toEntity());
	}

	@DeleteMapping(value="/{studentId}")
	public void delete(@PathVariable Long studentId) {

		studentService.delete(studentId);
	}


}
